tarde

+++ enrichisseur/database/migrations/2022_05_19_170657_add_category_id_terms_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::table('terms', function (Blueprint $table) {
             //
-            $table->unsignedBigInteger('category_id');
+            $table->unsignedInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
 
             Schema::enableForeignKeyConstraints();
